new day of cource

+++ section7/resources/item.py
@@ -1,6 +1,6 @@
 from flask import request
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
 from models.item import ItemModel
 
@@ -59,8 +59,11 @@ class Item(Resource):
         
         return item.json(), 201
 
-
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin requered!'}, 401 
         item = ItemModel.find_by_name(name)
         if item:
             item.delete()
